my_project/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
astor==0.8.1
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
Django==4.2.7
django-use-email-as-username==1.4.0
frozenlist==1.4.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==6.0.5
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
ultraimport==0.0.7
urllib3==2.2.1
yarl==1.9.4